Renaming API service
