+++ src/app/services/site-listing-facade.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { shareReplay } from 'rxjs/operators';
-import { SiteListingAPIService } from 'src/app/services/site-listing.service';
+import { SiteListingAPIService } from 'src/app/services/site-listing-api.service';
 import { AWSListing, AWSSiteListing } from 'src/app/services/sites.types';
 
 @Injectable({
